requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
cloudinary==1.33.0
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==3.2.19
django-allauth==0.54.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
oauthlib==3.2.2
pathspec==0.11.1
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
sqlparse==0.4.4
